Remove código e objetos redundantes
