+++ Kyuketsuki/Assets/Scripts/Fight Scripts/Battle Scripts/GameController.cs
@@ -24,7 +24,7 @@ public class GameController : MonoBehaviour
 
     private void Awake()
     {
-        Menu.instance.EmergencyDisableMenu();
+        GameMenu.instance.EmergencyDisableMenu();
         battleMenu = GameObject.Find("ActionMenu");
     }
     void Start()
