+++ Kyuketsuki/Assets/Scripts/GameMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 014193eb68998e04d86e611ff6ae62cf
+guid: ac5af02cb1cde204997f24077084b70b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
