+++ Kyuketsuki/Assets/Scripts/ItemButton.cs
@@ -23,11 +23,11 @@ public class ItemButton : MonoBehaviour
 
     public void Press()
     {
-        if (Menu.instance.GetGameMenuState() == MenuState.ItemScreen)
+        if (GameMenu.instance.GetGameMenuState() == MenuState.ItemScreen)
         {
             if(GameManager.instance.itemsHeld[buttonValue] != "")
             {
-                Menu.instance.SelectItem(GameManager.instance.GetItemDetails(GameManager.instance.itemsHeld[buttonValue]));
+                GameMenu.instance.SelectItem(GameManager.instance.GetItemDetails(GameManager.instance.itemsHeld[buttonValue]));
             }
         }
