--- Kyuketsuki/Assets/Scripts/Menu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac5af02cb1cde204997f24077084b70b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
